+++ R/modelStudio.R
@@ -51,6 +51,8 @@
 #' @param widget_id Use an explicit element ID for the widget (rather than an automatically generated one).
 #'  Useful e.g. when using \code{modelStudio} with Shiny.
 #'  See \href{https://modelstudio.drwhy.ai/articles/ms-perks-features.html#shiny-1}{\bold{vignette}}.
+#' @param license Path to the file containing the license (\code{con} parameter passed to \code{readLines()}).
+#'  It can be used e.g. to include the license for \code{explainer$data} as a comment in the source of \codE{.html} output file.
 #' @param telemetry The dashboard gathers useful, but not sensitive, information about how it is being used (e.g. computation length,
 #'  package version, dashboard dimensions). This is for the development purposes only and can be blocked by setting \code{telemetry} to \code{FALSE}.
 #' @param ... Other parameters.
@@ -180,6 +182,7 @@ modelStudio.explainer <- function(explainer,
                                   options = ms_options(),
                                   viewer = "external",
                                   widget_id = NULL,
+                                  license = NULL,
                                   telemetry = TRUE,
                                   max_vars = NULL,
                                   ...) {
@@ -456,6 +459,7 @@ modelStudio.explainer <- function(explainer,
                               obsCount = obs_count)
   }
 
+  if (!is.null(license)) options$license <- paste(readLines(license), collapse=" ")
   if (is.null(options$ms_title)) options$ms_title <- paste0("Interactive Studio for ", label, " Model")
   if (!is.null(options$ms_subtitle)) options$ms_margin_top <- options$ms_margin_top + 40